Fixes, HairStyle,Colors,Boots ect
Made a correction to the display player format and Aisling creation
flags.
hair colors and boots and names appear correctly.
still not sure how to make item names appear in equipment.
but will work on this later.

+++ Darkages.Server/Network/ClientFormats/ClientFormat04.cs
@@ -7,14 +7,14 @@
         public override byte Command => 0x04;
 
         public byte Gender { get; set; }
-        public short HairStyle { get; set; }
-        public short HairColor { get; set; }
+        public byte HairStyle { get; set; }
+        public byte HairColor { get; set; }
 
         public override void Serialize(NetworkPacketReader reader)
         {
             HairStyle = reader.ReadByte();
             Gender = reader.ReadByte();
-            HairColor = reader.ReadInt16();
+            HairColor = reader.ReadByte();
         }
 
         public override void Serialize(NetworkPacketWriter writer)
